requirements.txt
google-cloud-documentai==2.27.0
python-dotenv==1.0.1
PyPDF2==3.0.1
openai==1.35.2
PyYAML==6.0.1
dateparser==1.2.0
pandas==2.2.2